extract RadioButtons
